fixed std::string memory access error due to a corruption by curl api
