+++ liboai/include/core/netimpl.h
@@ -445,15 +445,16 @@ namespace liboai {
 				void PrepareDelete();
 				void PrepareDownload(std::ofstream& file);
 
-				void ParseResponseHeader(const std::string& headers, std::string* status_line, std::string* reason);
+				void ParseResponseHeader(const std::string& headers,
+					std::string* status_line, std::string* reason);
 
 				void SetOption(const components::Url& url);
 				void SetUrl(const components::Url& url);
 
 				void SetOption(const components::Body& body);
 				void SetBody(const components::Body& body);
-				void SetOption(components::Body&& body);
-				void SetBody(components::Body&& body);
+				//void SetOption(components::Body&& body);
+//				void SetBody(const components::Body& body);
 
 				void SetOption(const components::Multipart& multipart);
 				void SetMultipart(const components::Multipart& multipart);
@@ -508,6 +509,7 @@ namespace liboai {
 
 		template <class... _Options>
 		liboai::Response Post(_Options&&... options) {
+			 
 			Session session;
 			set_options(session, std::forward<_Options>(options)...);
 			return session.Post();
